increment5   - refactoring config   - updated tests
